Irix patches. Not fully tested. Thanks to Nekochan (http://www.nekochan.net) and Joerg Behrens for their support and their amazing Origin 3K.
git-svn-id: https://ssl.drqueue.org/svn/branches/0.64.x@1805 94d6b6b3-0ee6-0310-ab95-9b4161196ee2

+++ libdrqueue/request.c
@@ -2494,7 +2494,7 @@ void handle_r_r_jobdelblkhost (int sfd, struct database *wdb, int icomp, struct 
   uint32_t ijob;
   uint32_t ihost;
   struct blocked_host *nbh,*obh = NULL,*tnbh;
-  int nbhshmid;
+  int64_t nbhshmid;
   uint32_t i;
 
   log_auto(L_DEBUG,"Entering handle_r_r_jobdelblkhost");
@@ -2650,7 +2650,7 @@ void handle_r_r_jobblkhost (int sfd, struct database *wdb, int icomp, struct req
   uint32_t ijob;
   uint32_t ihost;
   struct blocked_host *nbh,*obh = NULL;
-  int nbhshmid;
+  int64_t nbhshmid;
   int i;
 
   log_auto(L_DEBUG,"Entering handle_r_r_jobblkhost");
